+++ config/keycloak-web.php
@@ -47,7 +47,7 @@ return [
     /**
      * Page to redirect after callback if there's no "intent"
      *
-     * @see Vizir\KeycloakWebGuard\Controllers\AuthController::callback()
+     * @see SlowCheetah\KeycloakWebGuard\Controllers\AuthController::callback()
      */
     'redirect_url' => '/admin',
 
@@ -58,7 +58,7 @@ return [
      *
      * The routes will receive the name "keycloak.{route}" and login/callback are required.
      * So, if you make it false, you shoul register a named 'keycloak.login' route and extend
-     * the Vizir\KeycloakWebGuard\Controllers\AuthController controller.
+     * the SlowCheetah\KeycloakWebGuard\Controllers\AuthController controller.
      */
     'routes' => [
         'login' => 'login',