+++ src/Middleware/KeycloakAuthenticated.php
@@ -1,11 +1,30 @@
 <?php
 
-namespace Vizir\KeycloakWebGuard\Middleware;
+namespace SlowCheetah\KeycloakWebGuard\Middleware;
 
+use Closure;
 use Illuminate\Auth\Middleware\Authenticate;
 
 class KeycloakAuthenticated extends Authenticate
 {
+    public function handle($request, Closure $next, ...$guards)
+    {
+        $excludedRoutes = [
+            route('keycloak.login', [],false),
+            route('keycloak.callback', [],false),
+        ];
+        $requestPath = '/' . $request->path();
+        foreach ($excludedRoutes as $excludedRoute) {
+            if ($requestPath === $excludedRoute) {
+                \Log::debug("Skipping {$excludedRoute} from auth check...");
+                return $next($request);
+            }
+        }
+
+        return parent::handle($request, $next, $guards);
+    }
+
+
     /**
      * Redirect user if it's not authenticated.
      *