+++ src/Middleware/KeycloakCan.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace Vizir\KeycloakWebGuard\Middleware;
+namespace SlowCheetah\KeycloakWebGuard\Middleware;
 
 use Closure;
 use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Support\Facades\Auth;
-use Vizir\KeycloakWebGuard\Exceptions\KeycloakCanException;
 
 class KeycloakCan extends KeycloakAuthenticated
 {